requirements.txt
protobuf==3.20.0
numpy==1.16.5
opencv-python==4.1.0.25
scipy==1.2.1
tensorflow==1.15.0
keras==2.2.4
git+https://github.com/divamgupta/image-segmentation-keras.git@f04852d3d51ac278e9e527d2eed78eddc5d56872
tqdm
h5py==2.10.0
pillow==6.0.0
